Final - (without Thomas feedback)

+++ SPAN/Container.h
@@ -1,3 +1,8 @@
+//Container.h
+//Jacob Schaupp
+//EECS 2510, Fall 2018
+//Contains the edge and vertex structures
+
 #pragma once
 #include "List.h"
 
@@ -13,7 +18,7 @@ struct edge
 struct vertex
 {
 	char key[1];
-	int path_length = 0;
+	double path_length = 0;
 	vertex *parent; //only used for prim's algorithm
 };
